refactor param_est to forecasting
