+++ aeon/forecasting/param_est/plugin.py
@@ -241,8 +241,8 @@ class PluginParamsForecaster(_DelegatedForecaster):
             `create_test_instance` uses the first (or only) dictionary in `params`
         """
         from aeon.forecasting.naive import NaiveForecaster
-        from aeon.param_est.fixed import FixedParams
-        from aeon.param_est.seasonality import SeasonalityACF
+        from aeon.forecasting.param_est.fixed import FixedParams
+        from aeon.forecasting.param_est.seasonality import SeasonalityACF
         from aeon.utils.validation._dependencies import _check_estimator_deps
 
         # use of dictionary to plug "foo" parameter into "sp", uses mock param_est
